pyproject.toml
[tool.poetry]
name = "my-practice"
version = "0.1.0"
description = ""
authors = ["Asgef <asgefes1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
flake8 = "^7.0.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
